Break tests into multiple files
