+++ test/helper.js
@@ -1,0 +1,4 @@
+global.assert = require('assert')
+global.example = require('../src/example')
+global.purified = require('../src')
+global.sinon = require('sinon')
