+++ test/mocha.opts
@@ -1,0 +1,1 @@
+--require ./test/helper.js
